Fixed some linux compiler errors and warnings and added the SIP Subscribe Clinet, Server and RefreshManager unit tests to the linux build.
git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@6359 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXtackLib/src/test/net/SipSubcribeServerTest.cpp
@@ -262,7 +262,7 @@ Voice-Message: 0/0 (0/0)\r\n";
        secondNotifyBody->getBytes(&notifyBodyBytes, &notifyBodySize);
        CPPUNIT_ASSERT(notifyBodyBytes);
        ASSERT_STR_EQUAL(mwiStateString, notifyBodyBytes);
-       CPPUNIT_ASSERT(notifyBodySize == strlen(mwiStateString));
+       CPPUNIT_ASSERT(notifyBodySize == (int)strlen(mwiStateString));
        CPPUNIT_ASSERT(notifyBodySize > 10);  // just to make sure both aren't null
        UtlString secondNotifyDialogHandle;
        secondNotify->getDialogHandle(secondNotifyDialogHandle);
